Correctly handling reference cycles between messages
Previously, a message referencing itself, such as for example:
```
message Foo {
    string name = 1;
    repeated Foo bar = 2;
}
```
or more generally messages depending on one another so as to form a dependency cycle
would result in an infinite loop when trying to generate JSON schemas
from them.

This patch fixes that by creating JSON schema references in those cases.

Added unit tests on the new behaviour.

Signed-off-by: Jean Rouge <jean@stripe.com>